Harden serial command parsing and buffer handling

+++ ENSABrew-100A.ino
@@ -560,7 +560,7 @@ boolean _newDataSerial = false;
 char receivedSerial[DATA_UTIL_SERIAL];
 
 static boolean _recvInProgress = false;
-static uint8_t idx = 0;
+static uint16_t idx = 0;
 volatile char dataChar;
 
 // ----------------------------------------------------------------------- //
